Remove if condition that is always false
Compiling with 'gcc (Debian 10.2.1-6) 10.2.1 20210110' and
-Wstring-compare gives the following warning:

In function ‘wc_locale_to_ces’,
    inlined from ‘wc_guess_locale_charset’ at libwc/charset.c:103:13,
    inlined from ‘wc_guess_locale_charset’ at libwc/charset.c:97:1:
libwc/charset.c:413:10: error: ‘strcmp’ of a string of length 8 and an array of size 8 evaluates to nonzero [-Werror=string-compare]
  413 |     if (!strcmp(buf, "japanese"))
      |          ^

If strcmp is always true, this condition is always false and was never
executed. It should be save to remove it.