Fix slow backward search in long lines
We need to search the whole line up to the position we are at to find
the first match from the end. If we have a match we know its end and can
skip past it. This makes searching very long strings way more efficient.

While there, refactor the whole function.

This fixes Debian BTS #1031413[0].

[0]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1031413