Fix gopher selector string
As can be seen in the gopher RFC[0] in the examples on pages 4 and 5 the
client is supposed to terminate the selection string with CR LF.

I only could track this bug back to the initial revision in the CVS
history, I would guess that it was present since the beginning of gopher
support. So, it seems it's not such a big deal, but I heard people
complain about it - just let's stick to the standard.

[0]: https://www.rfc-editor.org/rfc/rfc1436.html#section-2