Fix potential null dereference with gopher
guessContentType will return NULL if it can determine a ContentType.
This might happen e.g. if you insert a partial URL where the file
extension is missing.

Segfaulting is the worst of all options, so fall back to the default
ContentType of "image/png" in that case. Still leads to bad UX as the
process will continue without an obvious error besides the image not
being displayed.