Fix segfault with dump_source on compressed data
When one of -dump_source, -dump_both or -dump_extra is set and the
server returns compressed data the load function loadproc does not get
set because the content type for for compressed data is not handled in
the if-else block. This leads to reading of uninitialized memory when
the function is called through the pointer.

As of today I don't understand what should be done to dump in this case,
but a segfault is certainly the wrong result.

Set loadproc to NULL, which will lead to an error message, that the site
could not be loaded.