Drop upper limit on COLS
The parameter -cols has an undocumented arbitrary limit of 1024. If a
number greater than that is passed as an argument the value will be set
to 1024 instead.

I don't see a reason for this behavior. Drop that limit and let the user
decide.

This fixes Debian BTS #1076944 and GH issue #286.