Add password manager integration
This patch adds a command PASS and an option passwd_cmd. It is expected
that passwd_cmd takes the hostname part of the URL as an argument and
outputs user name and password on separate lines. It is tailored to be
used with password-store.

Let's make an example, I want to login to sr.ht and go to
https://meta.sr.ht/login. My password is stored with pass following the
convention, that the password is on the first line and my username on
the second. I have mapped PASS to the key 'p' and passwd_cmd is set to
'pass'. I place my cursor on the field 'Username' and press '2p'. w3m
will call 'pass meta.sr.ht', discard the first line and insert the
second line into the form. Then I jump to the password field and press
'p'. w3m call 'pass meta.sr.ht' again, but as I used no prefix, will
insert the first line into the form. Now I'm ready to submit the form
and login.

This is a first draft and I'm looking forward to hear any feedback on
this.