Add option exit_on_last
When exit_on_last is toggled on, w3m will close tabs and exit if the
user navigates back while on the last buffer in the stack. The use case
is to quit automatically when going back without the need for pressing a
different key or calling an additional command. This setting affects all
calls to backbf.

We change the descriptions of the other quit/close options to be more
accurate.