Revert "Fix segfault with dump_source on compressed data"
This reverts commit e933cf831f32fd4cf763c64aa60e51dbf025ec94.

This commit leads to segmentation faults when using -dump_source on
compressed data. It depends on another commit which is not committed
yet.

Also I can't reproduce the issue in any case and I currently don't see
what got 'fixed' here.  My guess is that I hit a case where bug in
loadBuffer was triggered and this worked around it by not calling it.