Enable re-ordering of buffers in the stack menu
I'm not happy with this implementation.

My original version did not use 'Buffer *prev'. It would check if we are at the
bottom of the stack with

    if (CurrentMenu->select == SelectMenu.nitem - 2) /* Bottom of menu */
       return (MENU_NOTHING);

but then GCC would warn about a potential null-dereference at

    if (cur == Firstbuf)
       Firstbuf = cur->nextBuffer;
    else
       prevBuffer(Firstbuf, cur)->nextBuffer = next;

because prevBuffer could return NULL if there were no previous buffer (which
can't happen because of the previous check).

Given the choice between losing a useful warning or wasting CPU cycles getting
prev even if we don't need it, I decided to do the unnecessary computation
because it's no big deal nowadays.