Replace STR_SIZE_MAX with STR_LEN_MAX
We are usually dealing with string length and not with the size a.k.a.
the length plus the terminating nul character or the allocated
area_size. The area_size is an implementation detail and should be as
little visible to the user as possible.

Use a STR_LEN_MAX to avoid dealing with of by errors.