Replace-borked-function-with-noop
The function insertself was the only one taking an argument in the
InputKeymap array. This lead to Undefined Behaviour. In commit 71c116c7
(Enable building with "modern C" build options) the argument was removed
on the caller site, which borked every use of insertself.

As inserting these CTRL keymaps does not make a lot of sense we just
disable this feature. As we also got no bug reports about insertself
being borked we can assume nobody does rely on this behaviour.